refactor(saved): extract hydrateSavedFromServer helper
SaveAuthPrompt and /login both ran the same Promise.allSettled block
fetching /api/session/saved-places and /api/session/saved-areas to
populate the session store. Move it next to putSavedList so there is
one place to update the hydration contract.
